feature: WIP sfc.gen.design -- add default slot when
component has text content or html content.

When the text content is empty (all whitespace) or starts with "optional"
the slot will be required: false, otherwise required: true.

However, the parser treats `<Card></Card>` and `<Card/>` the same, so
to get a slot you at least need a space or newline